# Protect against stack overflow caused by deep recursive calls