--- a/ARMeilleure/Optimizations.cs
+++ b/ARMeilleure/Optimizations.cs
@@ -6,8 +6,9 @@ namespace ARMeilleure
     {
         public static bool FastFP { get; set; } = true;
 
-        public static bool AllowLcqInFunctionTable  { get; set; } = true;
+        public static bool AllowLcqInFunctionTable { get; set; } = true;
         public static bool UseUnmanagedDispatchLoop { get; set; } = true;
+        public static bool EnableDeepCallRecursionProtection { get; set; } = true;
 
         public static bool UseSseIfAvailable       { get; set; } = true;
         public static bool UseSse2IfAvailable      { get; set; } = true;