On some games, the call stack might get deep, and this is problematic because it can consume a lot of stack space and eventually cause a overflow, which will cause the emulator to crash. This is possible because the CPU emulator implement guest function calls and function calls on the host too, as an optimization. In order to protect against this scenario, I have added a new `CallDepth` counter which keeps track of how deep it is on the call stack. Once it hits a given value, it will stop calling directly and return to the dispatcher. Doing so completely unwinds the stack.

This change might have a minor performance impact since we now need to do more work before each guest call, so I would recommend the following while testing:
- That there was no performance degradation.
- No noticeable increase in PPTC recompilation time.
- No large increase in the PPTC cache file size.

This fixes a crash that happens right after the first screen on Monster Hunter Rise with update 10.0.2. It should also fix a crash on the base game while looking at pictures taken with the camera on the game, but I did not check that one.