@liamwhite 2022-07-01
Something you might consider doing here is to avoid using the host stack at all and simply emulate the CPU RSB yourself -- it probably is a little bit slower, but it avoids many of the headaches associated with recompiling function calls with their own host stackframes.

In the JIT state, you could have something like

```cs
ulong[] rsb_pcs;
ulong[] rsb_host_addresses;
uint rsb_index;
```

On function call,
```c
rsb_index = (rsb_index + 1) % rsb_size;
rsb_pcs[rsb_index] = pc + 4;
rsb_host_addresses[rsb_index] = (native label);
goto function;
```

On function return,
```c
bool pop_hint = rsb_pcs[rsb_index] == lr;
rsb_index = (rsb_index - 1) % rsb_size;
if (pop_hint)
{
    goto rsb_host_addresses[rsb_index];
}
else
{
    goto dispatcher;
}
```

A similar method would apply for A32/T32, but have the additional complication of needing to deal with 16-bit instructions.