@gdkchan 2022-07-01
Something like this would require the ability to jump into the middle of a function. Right now the JIT does not support this and I'm not sure how it could be supported, as there's no place to store values that the function is using (it can save the values that directly match guest Arm registers on the `NativeContext`, but any other intermediate value on the function would be lost). Another option would be compiling a new function starting at the return point, where it would jump to on return, but then it would pollute the JIT cache with many smaller chuncks of the same function.