@FICTURE7 2022-07-01
I think there might be a correctness issue with this solution. If the return address passes the call guard in the caller then execution will continue in the caller and will not return back to the dispatcher. It seems like a rather slim possibility that would only happen in exotic handcrafted assembly.

If we're going with this solution, then the call depth check could be moved to the entry block of the target translation instead of before each call. This should reduce the amount of code we're emitting. Ideally we would emit the increase on entry and decrease on exit of each translation, but seems like a more involved change.

Side note, I have the suspicion the reason we're stack overflowing here is because of a pathological case I've in the hybrid allocator were it spills excessively and allocates a very large stack frame.