@gdkchan 2022-07-01
> I think there might be a correctness issue with this solution. If the return address passes the call guard in the caller then execution will continue in the caller and will not return back to the dispatcher. It seems like a rather slim possibility that would only happen in exotic handcrafted assembly.

Right, but I think it would still produce the correct results in this case, even if it doesn't return to the dispatcher, since it re-loads all guest register from the context after each call. Unless I'm missing something.

> If we're going with this solution, then the call depth check could be moved to the entry block of the target translation instead of before each call. This should reduce the amount of code we're emitting. Ideally we would emit the increase on entry and decrease on exit of each translation, but seems like a more involved change.

I don't think it would be much more complicated than this change, maybe I can give it a try. I guess doing that would also allow only enabling it for some functions (for example, the check is not needed on functions without any calls).

> Side note, I have the suspicion the reason we're stack overflowing here is because of a pathological case I've in the hybrid allocator were it spills excessively and allocates a very large stack frame.

That could be the case. It would be nice to have a register allocator that is as fast and produces better results.