@gdkchan 2022-07-01
> I don't think it would be much more complicated than this change, maybe I can give it a try. I guess doing that would also allow only enabling it for some functions (for example, the check is not needed on functions without any calls).

I forgot one important detail, there are also tail calls, it shouldn't increase the call depth value for those. If they were added to the function entry, it would be tricky to tell, I guess it would need to store some value on the context to indicate that or pass it as an argument to the function.