@FICTURE7 2022-07-02
> Right, but I think it would still produce the correct results in this case, even if it doesn't return to the dispatcher, since it re-loads all guest register from the context after each call. Unless I'm missing something.

Yeah you're right. After further thought it does not appear to be a problem.

> I forgot one important detail, there are also tail calls, it shouldn't increase the call depth value for those. If they were added to the function entry, it would be tricky to tell, I guess it would need to store some value on the context to indicate that or pass it as an argument to the function.

I think we can decrement the call depth just before a tailcall happens, it should avoid that issue.

> That could be the case. It would be nice to have a register allocator that is as fast and produces better results.

IIRC #2695 is pretty much ready; it is slower than the current allocator (2-5%), but the CQ gains were very good in certain cases. I will see if I can push it to the finish line.