@gdkchan 2022-07-10
> > I forgot one important detail, there are also tail calls, it shouldn't increase the call depth value for those. If they were added to the function entry, it would be tricky to tell, I guess it would need to store some value on the context to indicate that or pass it as an argument to the function.
> 
> I think we can decrement the call depth just before a tailcall happens, it should avoid that issue.

I changed to increment the call depth at the start of the function, rather than before call. Seems to work. I also took the opportunity to get more information about the problematic calls on Monster Hunter Rise. First, I can confirm that it appears to be indeed a recursive call. There are 2 functions being called in alternation on the call stack:
```
0xE2003C0
0xE200338
0xE2003C0
0xE200338
```
(Those are guest addresses).
The stack overflows after 1596 calls.