@MutantAura 2022-08-15
Performance: Couldn't notice anything user-side in any games. Seemed to be within variance.

CPU tests in bytemark (Master):
![image](https://user-images.githubusercontent.com/44103205/184670910-6ec44fc0-b307-4289-be20-12bfe38911e6.png)

CPU tests in bytemark (PR):
![image](https://user-images.githubusercontent.com/44103205/184670971-593bac2d-e06f-403d-bab3-fefa9729fd11.png)

PPTC recompilation tests and filesizes.
1. Run game in master.
2. Run game in PR (rebased) to recompile PPTC and time.
3. Run game in master again to recompile PPTC and time.

Super Mario Odyssey 
Master: 4.25s (63.6 MB)
PR: 4.15s (67.0 MB)

Breath of the Wild
Master: 22.44s (258 MB)
PR: 22.75s (265 MB)

Super Smash Bros: Ultimate
Master: 21.05s (183 MB)
PR: 22.18s (195 MB)

Dragon Quest XI
Master: 17.85s (151 MB)
PR: 18.31s (161 MB)